# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/communities": {
      "get": {
        "operationId": "CommunityController_communities",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "post": {
        "operationId": "CommunityController_createCommunity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/communities/{id}": {
      "get": {
        "operationId": "CommunityController_community",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "patch": {
        "operationId": "CommunityController_updateCommunity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "delete": {
        "operationId": "CommunityController_deleteCommunity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/towers": {
      "get": {
        "operationId": "CommunityController_towers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "post": {
        "operationId": "CommunityController_createTower",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTowerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/towers/{id}": {
      "get": {
        "operationId": "CommunityController_tower",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "patch": {
        "operationId": "CommunityController_updateTower",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTowerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "delete": {
        "operationId": "CommunityController_deleteTower",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/floors": {
      "get": {
        "operationId": "CommunityController_floors",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "post": {
        "operationId": "CommunityController_createFloor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFloorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/floors/{id}": {
      "get": {
        "operationId": "CommunityController_floor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "patch": {
        "operationId": "CommunityController_updateFloor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFloorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "delete": {
        "operationId": "CommunityController_deleteFloor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/apartments": {
      "get": {
        "operationId": "CommunityController_apartments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "post": {
        "operationId": "CommunityController_createApartment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/apartments/{id}": {
      "get": {
        "operationId": "CommunityController_apartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "patch": {
        "operationId": "CommunityController_updateApartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      },
      "delete": {
        "operationId": "CommunityController_deleteApartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/users/{id}/resident-apartment": {
      "patch": {
        "operationId": "CommunityController_associateResident",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssociateApartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/users/{id}/representative-tower": {
      "patch": {
        "operationId": "CommunityController_associateRepresentative",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssociateTowerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/users/me/hierarchy": {
      "get": {
        "operationId": "CommunityController_resolveOwnHierarchy",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
    "/users/{id}/resident-hierarchy": {
      "get": {
        "operationId": "CommunityController_resolveResidentHierarchy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "community hierarchy"
        ]
      }
    },
...
  },
  "info": {
    "title": "Urbanity API",
    "description": "In-memory NestJS CRUD API with JWT-backed role-based access control.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCommunityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "address"
        ]
      },
      "UpdateCommunityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateTowerDto": {
        "type": "object",
        "properties": {
          "communityId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "communityId",
          "name",
          "code"
        ]
      },
      "UpdateTowerDto": {
        "type": "object",
        "properties": {
          "communityId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateFloorDto": {
        "type": "object",
        "properties": {
          "towerId": {
            "type": "string"
          },
          "floorNumber": {
            "type": "number"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "towerId",
          "floorNumber",
          "label"
        ]
      },
      "UpdateFloorDto": {
        "type": "object",
        "properties": {
          "towerId": {
            "type": "string"
          },
          "floorNumber": {
            "type": "number"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "CreateApartmentDto": {
        "type": "object",
        "properties": {
          "floorId": {
            "type": "string"
          },
          "apartmentNumber": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "floorId",
          "apartmentNumber",
          "label"
        ]
      },
      "UpdateApartmentDto": {
        "type": "object",
        "properties": {
          "floorId": {
            "type": "string"
          },
          "apartmentNumber": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "AssociateApartmentDto": {
        "type": "object",
        "properties": {
          "apartmentId": {
            "type": "string"
          }
        },
        "required": [
          "apartmentId"
        ]
      },
      "AssociateTowerDto": {
        "type": "object",
        "properties": {
          "towerId": {
            "type": "string"
          }
        },
        "required": [
          "towerId"
        ]
      },
...
    }
  }
}